// service/src/interfaces/entity/IRedditHotArtificialRaw.ts
export interface Child {
    kind: string;
    data: Data2;
}

export interface Data {
    after?:string;
    dist: number;
    modhash: string;
    geo_filter?: any;
    children: Child[];
    before?: string;
}

export interface IRedditHotArtificialRaw {
    kind: string;
    data: Data;
}